+++ src/lapidary/runtime/__init__.py
@@ -4,6 +4,9 @@ __all__ = (
     'NamedAuth',
     'ParamStyle',
     'RequestBody',
+    'ResponseEnvelope',
+    'ResponseHeader',
+    'ResponseBody',
     'Responses',
     'SecurityRequirements',
     'delete',
@@ -22,8 +25,8 @@ __all__ = (
 from .client_base import ClientBase
 from .model import ModelBase
 from .model.encode_param import ParamStyle
-from .model.params import RequestBody
-from .model.response_map import Responses
+from .model.params import Body as RequestBody
+from .model.response import Body as ResponseBody, Header as ResponseHeader, ResponseEnvelope, Responses
 from .operation import delete, get, head, patch, post, put, trace
 from .param import Cookie, Header, Path, Query
 from .types_ import NamedAuth, SecurityRequirements